1. Fixed compilation error due to Phonon inclusion. This was due to wrong 	platform selector in .pro file 2. Phonon engine is now restricted to Linux uisng .pri file, no #ifdef trick 3. Fixed some warnings

+++ src/desktop/widgets/SuperSlider.cpp
@@ -226,7 +226,7 @@ void SuperSlider::setMaxValue( qint64 maxValue )
 
 void SuperSlider::setCurrentValue( qint64 curValue )
 {
-    if( curValue <= m_maxVal && m_curentVal != curValue ) {
+    if( curValue <= m_maxVal && m_curentVal != (quint64 ) curValue ) {
         m_curentVal = curValue < 0 ? 0
                                    : ( curValue > m_maxVal ? m_maxVal
                                                            : curValue );
